--- src/components/Cards/NetAssetTag.tsx
@@ -1,20 +1,0 @@
-import type { FC } from 'react'
-
-import { assetAmountToString } from '../../utils/helpers'
-
-interface TagProps {
-  assetsum: number
-}
-
-const NetAssetTag: FC<TagProps> = ({ assetsum }) => {
-  return (
-    <div className="flex min-w-[150px] max-w-fit flex-col items-center justify-center gap-2 rounded-xl bg-lightBlue px-4 py-2">
-      <div className="flex text-center">{'net asset value'}</div>
-      <div className="header-2 flex pb-1 text-cobalt">
-        {assetAmountToString(assetsum)}
-      </div>
-    </div>
-  )
-}
-
-export default NetAssetTag
